requirements.txt
Babel==0.9.6
Flask==0.9
Flask-Babel==0.8
Flask-Login==0.2.9
Flask-Mail==0.8.2
Flask-SQLAlchemy==0.16
Flask-WTF==0.8.3
Flask-HTTPAuth
Flask-Restful
Flask-Principal
Flask-Script
Flask-Security
Flask-Uploads
Flask-Storage
git+git://github.com/miguelgrinberg/Flask-WhooshAlchemy
Jinja2==2.6
MySQL-python==1.2.4
SQLAlchemy==0.7.9
Tempita==0.5.1
WTForms==1.0.3
Werkzeug==0.8.3
Whoosh==2.4.1
blinker==1.2
coverage==3.6
decorator==3.4.0
flup==1.0.3.dev-20110405
guess-language==0.2
gunicorn==0.17.2
pytz==2013b
speaklater==1.3
sqlalchemy-migrate==0.7.2
passlib
itsdangerous
rauth
facebook-sdk
pymongo
#pycrypto
#Flask-OpenID==1.1.1
#psycopg2
#python-openid==2.2.5
